Remove useless struct: mbfl_filt_tl_jisx0201_jisx0208_param
This struct contained an unused 'next_filter' field. The only field which was used
was the integral 'mode'. Since only one field in the struct was used, it is better
to inline it and just store the integer directly where it is needed.

Further, while an instance of this struct was included in the struct
mbfl_filt_conv_wchar_cp50220_ctx, it was never actually used there. It was just
taking up memory for nothing.

+++ ext/mbstring/libmbfl/mbfl/mbfilter.h
@@ -85,6 +85,7 @@
 #ifndef MBFL_MBFILTER_H
 #define MBFL_MBFILTER_H
 
+#include <stdint.h>
 #include "zend.h"
 
 #include "mbfl_defs.h"
@@ -294,9 +295,8 @@ MBFLAPI extern mbfl_string *
 mbfl_html_numeric_entity_decode(mbfl_string *string, mbfl_string *result, int *convmap, int mapsize);
 
 /*
- * convert of harfwidth and fullwidth for japanese
+ * conversion of halfwidth and fullwidth characters for Japanese
  */
-MBFLAPI extern mbfl_string *
-mbfl_ja_jp_hantozen(mbfl_string *string, mbfl_string *result, int mode);
+MBFLAPI extern mbfl_string* mbfl_ja_jp_hantozen(mbfl_string *string, mbfl_string *result, intptr_t mode);
 
 #endif	/* MBFL_MBFILTER_H */
